+++ internal/command/jsonstate/state_test.go
@@ -5,6 +5,7 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/google/go-cmp/cmp"
 	"github.com/hashicorp/terraform/internal/addrs"
 	"github.com/hashicorp/terraform/internal/configs/configschema"
 	"github.com/hashicorp/terraform/internal/states"
@@ -180,7 +181,7 @@ func TestMarshalResources(t *testing.T) {
 			},
 			testSchemas(),
 			[]resource{
-				resource{
+				{
 					Address:      "test_thing.bar",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -191,6 +192,7 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
 			},
 			false,
@@ -250,7 +252,7 @@ func TestMarshalResources(t *testing.T) {
 			},
 			testSchemas(),
 			[]resource{
-				resource{
+				{
 					Address:      "test_thing.bar[0]",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -261,6 +263,7 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
 			},
 			false,
@@ -291,7 +294,7 @@ func TestMarshalResources(t *testing.T) {
 			},
 			testSchemas(),
 			[]resource{
-				resource{
+				{
 					Address:      "test_thing.bar[\"rockhopper\"]",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -302,6 +305,7 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
 			},
 			false,
@@ -319,7 +323,7 @@ func TestMarshalResources(t *testing.T) {
 					Instances: map[addrs.InstanceKey]*states.ResourceInstance{
 						addrs.NoKey: {
 							Deposed: map[states.DeposedKey]*states.ResourceInstanceObjectSrc{
-								states.DeposedKey(deposedKey): &states.ResourceInstanceObjectSrc{
+								states.DeposedKey(deposedKey): {
 									Status:    states.ObjectReady,
 									AttrsJSON: []byte(`{"woozles":"confuzles"}`),
 								},
@@ -334,7 +338,7 @@ func TestMarshalResources(t *testing.T) {
 			},
 			testSchemas(),
 			[]resource{
-				resource{
+				{
 					Address:      "test_thing.bar",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -346,6 +350,7 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
 			},
 			false,
@@ -363,7 +368,7 @@ func TestMarshalResources(t *testing.T) {
 					Instances: map[addrs.InstanceKey]*states.ResourceInstance{
 						addrs.NoKey: {
 							Deposed: map[states.DeposedKey]*states.ResourceInstanceObjectSrc{
-								states.DeposedKey(deposedKey): &states.ResourceInstanceObjectSrc{
+								states.DeposedKey(deposedKey): {
 									Status:    states.ObjectReady,
 									AttrsJSON: []byte(`{"woozles":"confuzles"}`),
 								},
@@ -382,7 +387,7 @@ func TestMarshalResources(t *testing.T) {
 			},
 			testSchemas(),
 			[]resource{
-				resource{
+				{
 					Address:      "test_thing.bar",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -393,8 +398,9 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
-				resource{
+				{
 					Address:      "test_thing.bar",
 					Mode:         "managed",
 					Type:         "test_thing",
@@ -406,6 +412,7 @@ func TestMarshalResources(t *testing.T) {
 						"foozles": json.RawMessage(`null`),
 						"woozles": json.RawMessage(`"confuzles"`),
 					},
+					SensitiveValues: json.RawMessage("{}"),
 				},
 			},
 			false,
@@ -423,10 +430,12 @@ func TestMarshalResources(t *testing.T) {
 			} else if err != nil {
 				t.Fatalf("unexpected error: %s", err)
 			}
-			eq := reflect.DeepEqual(got, test.Want)
-			if !eq {
-				t.Fatalf("wrong result:\nGot: %#v\nWant: %#v\n", got, test.Want)
+
+			diff := cmp.Diff(got, test.Want)
+			if diff != "" {
+				t.Fatalf("wrong result: %s\n", diff)
 			}
+
 		})
 	}
 }
@@ -629,12 +638,12 @@ func TestMarshalModules_parent_no_resources(t *testing.T) {
 func testSchemas() *terraform.Schemas {
 	return &terraform.Schemas{
 		Providers: map[addrs.Provider]*terraform.ProviderSchema{
-			addrs.NewDefaultProvider("test"): &terraform.ProviderSchema{
+			addrs.NewDefaultProvider("test"): {
 				ResourceTypes: map[string]*configschema.Block{
 					"test_thing": {
 						Attributes: map[string]*configschema.Attribute{
 							"woozles": {Type: cty.String, Optional: true, Computed: true},
-							"foozles": {Type: cty.String, Optional: true},
+							"foozles": {Type: cty.String, Optional: true, Sensitive: true},
 						},
 					},
 					"test_instance": {
@@ -647,5 +656,218 @@ func testSchemas() *terraform.Schemas {
 				},
 			},
 		},
+	}
+}
+
+func TestSensitiveAsBool(t *testing.T) {
+	sensitive := "sensitive"
+	tests := []struct {
+		Input cty.Value
+		Want  cty.Value
+	}{
+		{
+			cty.StringVal("hello"),
+			cty.False,
+		},
+		{
+			cty.NullVal(cty.String),
+			cty.False,
+		},
+		{
+			cty.StringVal("hello").Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.NullVal(cty.String).Mark(sensitive),
+			cty.True,
+		},
+
+		{
+			cty.NullVal(cty.DynamicPseudoType).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.NullVal(cty.Object(map[string]cty.Type{"test": cty.String})),
+			cty.False,
+		},
+		{
+			cty.NullVal(cty.Object(map[string]cty.Type{"test": cty.String})).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.DynamicVal,
+			cty.False,
+		},
+		{
+			cty.DynamicVal.Mark(sensitive),
+			cty.True,
+		},
+
+		{
+			cty.ListValEmpty(cty.String),
+			cty.EmptyTupleVal,
+		},
+		{
+			cty.ListValEmpty(cty.String).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.StringVal("hello"),
+				cty.StringVal("friend").Mark(sensitive),
+			}),
+			cty.TupleVal([]cty.Value{
+				cty.False,
+				cty.True,
+			}),
+		},
+		{
+			cty.SetValEmpty(cty.String),
+			cty.EmptyTupleVal,
+		},
+		{
+			cty.SetValEmpty(cty.String).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.SetVal([]cty.Value{cty.StringVal("hello")}),
+			cty.TupleVal([]cty.Value{cty.False}),
+		},
+		{
+			cty.SetVal([]cty.Value{cty.StringVal("hello").Mark(sensitive)}),
+			cty.True,
+		},
+		{
+			cty.EmptyTupleVal.Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.TupleVal([]cty.Value{
+				cty.StringVal("hello"),
+				cty.StringVal("friend").Mark(sensitive),
+			}),
+			cty.TupleVal([]cty.Value{
+				cty.False,
+				cty.True,
+			}),
+		},
+		{
+			cty.MapValEmpty(cty.String),
+			cty.EmptyObjectVal,
+		},
+		{
+			cty.MapValEmpty(cty.String).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.MapVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse"),
+			}),
+			cty.EmptyObjectVal,
+		},
+		{
+			cty.MapVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse").Mark(sensitive),
+			}),
+			cty.ObjectVal(map[string]cty.Value{
+				"animal": cty.True,
+			}),
+		},
+		{
+			cty.MapVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse").Mark(sensitive),
+			}).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.EmptyObjectVal,
+			cty.EmptyObjectVal,
+		},
+		{
+			cty.ObjectVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse"),
+			}),
+			cty.EmptyObjectVal,
+		},
+		{
+			cty.ObjectVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse").Mark(sensitive),
+			}),
+			cty.ObjectVal(map[string]cty.Value{
+				"animal": cty.True,
+			}),
+		},
+		{
+			cty.ObjectVal(map[string]cty.Value{
+				"greeting": cty.StringVal("hello"),
+				"animal":   cty.StringVal("horse").Mark(sensitive),
+			}).Mark(sensitive),
+			cty.True,
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.UnknownVal(cty.String),
+				}),
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.StringVal("known").Mark(sensitive),
+				}),
+			}),
+			cty.TupleVal([]cty.Value{
+				cty.EmptyObjectVal,
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.True,
+				}),
+			}),
+		},
+		{
+			cty.ListVal([]cty.Value{
+				cty.MapValEmpty(cty.String),
+				cty.MapVal(map[string]cty.Value{
+					"a": cty.StringVal("known").Mark(sensitive),
+				}),
+				cty.MapVal(map[string]cty.Value{
+					"a": cty.UnknownVal(cty.String),
+				}),
+			}),
+			cty.TupleVal([]cty.Value{
+				cty.EmptyObjectVal,
+				cty.ObjectVal(map[string]cty.Value{
+					"a": cty.True,
+				}),
+				cty.EmptyObjectVal,
+			}),
+		},
+		{
+			cty.ObjectVal(map[string]cty.Value{
+				"list":   cty.UnknownVal(cty.List(cty.String)),
+				"set":    cty.UnknownVal(cty.Set(cty.Bool)),
+				"tuple":  cty.UnknownVal(cty.Tuple([]cty.Type{cty.String, cty.Number})),
+				"map":    cty.UnknownVal(cty.Map(cty.String)),
+				"object": cty.UnknownVal(cty.Object(map[string]cty.Type{"a": cty.String})),
+			}),
+			cty.ObjectVal(map[string]cty.Value{
+				"list":   cty.EmptyTupleVal,
+				"set":    cty.EmptyTupleVal,
+				"tuple":  cty.EmptyTupleVal,
+				"map":    cty.EmptyObjectVal,
+				"object": cty.EmptyObjectVal,
+			}),
+		},
+	}
+
+	for _, test := range tests {
+		got := SensitiveAsBool(test.Input)
+		if !reflect.DeepEqual(got, test.Want) {
+			t.Errorf(
+				"wrong result\ninput: %#v\ngot:   %#v\nwant:  %#v",
+				test.Input, got, test.Want,
+			)
+		}
 	}
 }
