+++ internal/command/testdata/show-json-state/basic/output.json
@@ -1,5 +1,5 @@
 {
-    "format_version": "0.1",
+    "format_version": "0.2",
     "terraform_version": "0.12.0",
     "values": {
         "root_module": {
@@ -15,7 +15,8 @@
                     "values": {
                         "ami": null,
                         "id": "621124146446964903"
-                    }
+                    },
+                    "sensitive_values": {}
                 },
                 {
                     "address": "test_instance.example[1]",
@@ -28,7 +29,8 @@
                     "values": {
                         "ami": null,
                         "id": "4330206298367988603"
-                    }
+                    },
+                    "sensitive_values": {}
                 }
             ]
         }
