+++ internal/command/testdata/show-json-state/empty/output.json
@@ -1,3 +1,3 @@
 {
-    "format_version": "0.1"
+    "format_version": "0.2"
 }
