+++ internal/command/testdata/show-json-state/modules/output.json
@@ -1,5 +1,5 @@
 {
-    "format_version": "0.1",
+    "format_version": "0.2",
     "terraform_version": "0.12.0",
     "values": {
         "outputs": {
@@ -22,7 +22,8 @@
                             "values": {
                                 "ami": "bar-var",
                                 "id": null
-                            }
+                            },
+                            "sensitive_values": {}
                         }
                     ],
                     "address": "module.module_test_bar"
@@ -40,7 +41,8 @@
                             "values": {
                                 "ami": "foo-var",
                                 "id": null
-                            }
+                            },
+                            "sensitive_values": {}
                         }
                     ],
                     "address": "module.module_test_foo"
