+++ internal/command/testdata/show-json/basic-delete/output.json
@@ -1,5 +1,5 @@
 {
-    "format_version": "0.1",
+    "format_version": "0.2",
     "variables": {
         "test_var": {
             "value": "bar"
@@ -24,7 +24,8 @@
                     "values": {
                         "ami": "bar",
                         "id": "placeholder"
-                    }
+                    },
+                    "sensitive_values": {}
                 }
             ]
         }
@@ -87,7 +88,7 @@
         }
     },
     "prior_state": {
-        "format_version": "0.1",
+        "format_version": "0.2",
         "values": {
             "outputs": {
                 "test": {
@@ -107,7 +108,8 @@
                         "values": {
                             "ami": "foo",
                             "id": "placeholder"
-                        }
+                        },
+                        "sensitive_values": {}
                     },
                     {
                         "address": "test_instance.test-delete",
@@ -119,7 +121,8 @@
                         "values": {
                             "ami": "foo",
                             "id": "placeholder"
-                        }
+                        },
+                        "sensitive_values": {}
                     }
                 ]
             }
