+++ internal/command/testdata/show-json/module-depends-on/output.json
@@ -1,5 +1,5 @@
 {
-    "format_version": "0.1",
+    "format_version": "0.2",
     "terraform_version": "0.13.1-dev",
     "planned_values": {
         "root_module": {
@@ -13,7 +13,8 @@
                     "schema_version": 0,
                     "values": {
                         "ami": "foo-bar"
-                    }
+                    },
+                    "sensitive_values": {}
                 }
             ]
         }
