+++ internal/command/testdata/show-json/modules/output.json
@@ -1,5 +1,5 @@
 {
-    "format_version": "0.1",
+    "format_version": "0.2",
     "planned_values": {
         "outputs": {
             "test": {
@@ -20,7 +20,8 @@
                             "schema_version": 0,
                             "values": {
                                 "ami": "bar-var"
-                            }
+                            },
+                            "sensitive_values": {}
                         }
                     ],
                     "address": "module.module_test_bar"
@@ -37,7 +38,8 @@
                             "schema_version": 0,
                             "values": {
                                 "ami": "baz"
-                            }
+                            },
+                            "sensitive_values": {}
                         },
                         {
                             "address": "module.module_test_foo.test_instance.test[1]",
@@ -49,7 +51,8 @@
                             "schema_version": 0,
                             "values": {
                                 "ami": "baz"
-                            }
+                            },
+                            "sensitive_values": {}
                         },
                         {
                             "address": "module.module_test_foo.test_instance.test[2]",
@@ -61,7 +64,8 @@
                             "schema_version": 0,
                             "values": {
                                 "ami": "baz"
-                            }
+                            },
+                            "sensitive_values": {}
                         }
                     ],
                     "address": "module.module_test_foo"
@@ -70,7 +74,7 @@
         }
     },
     "prior_state": {
-        "format_version": "0.1",
+        "format_version": "0.2",
         "values": {
             "outputs": {
                 "test": {
