+++ internal/command/testdata/show-json/nested-modules/output.json
@@ -1,5 +1,5 @@
 {
-  "format_version": "0.1",
+  "format_version": "0.2",
   "planned_values": {
     "root_module": {
       "child_modules": [
@@ -17,7 +17,8 @@
                   "schema_version": 0,
                   "values": {
                     "ami": "bar-var"
-                  }
+                  },
+                  "sensitive_values": {}
                 }
               ],
               "address": "module.my_module.module.more"
